Updated models, and API with middleware

+++ src/app/Http/Controllers/TicketController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Ticket;
 use App\User;
 use App\Http\Requests\TicketFormRequest;
+use Mockery\Exception;
 
 class TicketController extends Controller
 {
@@ -27,6 +28,7 @@ class TicketController extends Controller
             'type' => $allRequest['type'],
             'desc' => $allRequest['description'],
             'user_id' => $newUser->id,
+            'staff_id' => null,
             'completed' => false]);
         $newTicket->save();
 
@@ -56,14 +58,39 @@ class TicketController extends Controller
 
     //Updates ticket with corresponding id
     public function update(Request $request, $id) {
+        try {
+            $ticket = Ticket::find($id);
+            $ticket->name = $request->name;
+            $ticket->details = $request->details;
 
+            $saved = $ticket->save();
+
+            if(!$saved){
+                return array("status" => "error");
+            }
+        }
+        catch(Exception $e) {
+            return array("status" => "error");
+        }
+
+        return array("status" => "success");
     }
 
     //Deletes ticket with corresponding id
-    public function delete(Request $request, $id) {
-        $ticket = Ticket::findOrFail($id);
-        $ticket->delete();
+    public function delete($id) {
+        try {
+            $ticket = Ticket::find($id);
+            if($ticket != null) {
+                $ticket->delete();
+            } else {
+                return array("status" => "error");
+            }
 
-        return 204;
+            return array("status" => "success");
+        } catch(Exception $e) {
+            return array("status" => "error");
+        }
+
+        return array("status" => "success");
     }
 }
